Installation brought up some issues with dictionaries-common. Fixed this issue and included ansible
Conflicts:
	Vagrantfile

+++ Vagrantfile
@@ -17,7 +17,9 @@ Vagrant.configure(2) do |config|
     sudo echo "LANGUAGE=en_US.UTF-8" >> /etc/environment
     sudo echo "LC_ALL=en_US.UTF-8" >> /etc/environment
     sudo echo "LC_CTYPE=en_US.UTF-8" >> /etc/environment
+    export DEBIAN_FRONTEND=noninteractive
     sudo add-apt-repository -y ppa:webupd8team/java
+    sudo apt-add-repository ppa:ansible/ansible
     sudo apt-get update
     sudo apt-get -y upgrade
     echo debconf shared/accepted-oracle-license-v1-1 select true | sudo debconf-set-selections 
@@ -27,13 +29,26 @@ Vagrant.configure(2) do |config|
     sudo apt-get install -y xfce4 virtualbox-guest-dkms virtualbox-guest-utils virtualbox-guest-x11
     sudo apt-get install gnome-icon-theme-full tango-icon-theme
     sudo echo "allowed_users=anybody" > /etc/X11/Xwrapper.config
-  
-    sudo wget -O /opt/eclipse-committers-oxygen-3-linux-gtk-x86_64.tar.gz http://ftp.fau.de/eclipse/technology/epp/downloads/release/oxygen/3/eclipse-committers-oxygen-3-linux-gtk-x86_64.tar.gz
-    cd /opt/ && sudo tar -zxvf eclipse-committers-oxygen-3-linux-gtk-x86_64.tar.gz
+
+    sudo /usr/share/debconf/fix_db.pl
+    sudo sudo apt-get -y upgrade
+
+    if ! [ -f /opt/eclipse-committers-oxygen-3-linux-gtk-x86_64.tar.gz ]; then
+        sudo wget -O /opt/eclipse-committers-oxygen-3-linux-gtk-x86_64.tar.gz http://ftp.fau.de/eclipse/technology/epp/downloads/release/oxygen/3/eclipse-committers-oxygen-3-linux-gtk-x86_64.tar.gz
+        cd /opt/ && sudo tar -zxvf eclipse-committers-oxygen-3-linux-gtk-x86_64.tar.gz
+    fi
 
     curl -sL https://deb.nodesource.com/setup_9.x | sudo -E bash -
     sudo apt-get install -y nodejs
     echo "installed node js"
+
+    sudo apt-get -q -y install software-properties-common
+
+    sudo apt-get -q -y install ansible python-pip python-dev libffi-dev libssl-dev
+
+    sudo pip install pytz shade
+
+
 
   SHELL
